CLS-13: Simplify: Move CommandLineWrapper to the ExecutionControl directory
